modif

+++ src/app/Enseignant/add-enseignant-to-groupe/add-enseignant-to-groupe.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { Groupe } from 'src/app/Groupe/groupe';
 import { GroupeService } from 'src/app/Groupe/groupe.service';
@@ -11,6 +12,8 @@ import { EnseignantService } from '../enseignant.service';
   styleUrls: ['./add-enseignant-to-groupe.component.css']
 })
 export class AddEnseignantToGroupeComponent implements OnInit {
+  enseignantGroupeFormgroup!:FormGroup;
+  submitted:boolean=false;
   groupeId!:number;
   mail!:string;
   numeroGroupe!:string;
@@ -19,15 +22,20 @@ export class AddEnseignantToGroupeComponent implements OnInit {
   groupes: Groupe[] = [];
   enseignants: Enseignant[] = [];
   constructor(private enseignantService:EnseignantService,
-    private router:Router,private groupeService:GroupeService) { }
+    private router:Router,private groupeService:GroupeService,private fb:FormBuilder) { }
 
   ngOnInit(): void {
+    this.enseignantGroupeFormgroup=this.fb.group({
+      egEnseignant:["",Validators.required],
+      egGroupe:["",Validators.required]
+    });
     this.getEnseigants();
     this.getGroupes();
   }
 
   onSubmit(){
-   
+    this.submitted=true;
+    if(this.enseignantGroupeFormgroup.invalid) return;
     this.enseignantService.addEnseignantToGroupe(this.mail,this.numeroGroupe).subscribe(data => {
       console.log(data);
       alert("Enseignant ajouté au groupe avec succés !");
@@ -36,11 +44,13 @@ export class AddEnseignantToGroupeComponent implements OnInit {
     
   }
   onRemove(){
+    this.submitted=true;
+    if(this.enseignantGroupeFormgroup.invalid) return;
     let conf=confirm("Etes-vous sure ?")
     if(conf){
     this.enseignantService.deleteEnseignantToGroupe(this.mail,this.numeroGroupe).subscribe(data => {
       console.log(data);
-     
+      alert("Enseignant retiré du groupe avec succés !");
     });
     
   }
